+++ packages/xchain-bitcoin/src/client.ts
@@ -42,7 +42,7 @@ class Client extends UTXOClient {
    * @param {BitcoinClientParams} params
    */
   constructor({
-    network = Network.Testnet,
+    network = Network.Mainnet,
     feeBounds = {
       lower: LOWER_FEE_BOUND,
       upper: UPPER_FEE_BOUND,